+++ CHANGELOG.md
@@ -39,6 +39,7 @@ Next, enable the PSPDFKit.xcconfig project configuration. Here's a screenshot: h
 *  The PSPDFKit image annotation now properly displays and animates animated GIFs.
 *  A long-press on a PDF link annotation now shows the URL/Document/Page target in a popover.
 *  The PSPDFScrobbleBar now uses the small height style on iPhone/Landscape.
+*  PSPDFWebViewController now supports printing.
 *  Fixes a problem in the annotation parser where some named page links failed to resolve properly.
 *  Fixes a problem where the Cancel button of the additionalBarButtonItem menu wasn't fully touch-able on iPhone.
 *  Fixes a problem with PSPDFBarButtonItem image updating.